Split App.jsx into component architecture (Issue #239)
Refactor-only change: no behavior modifications.

App.jsx: 3176 lines → 38 lines (slim shell with routing).

New files:
- hooks/useAuth.js: auth state + bootstrap orchestration
- services/api.js: authHeaders, getAccessToken, describeFetchFailure
- services/routing.js: hash-based route utilities
- components/AppShell.jsx: signed-in shell (topbar, rail, modals)
- components/MarketingShell.jsx: signed-out landing page
- components/MigrationShell.jsx: /migration page with upload
- components/common/: Icon, modals, HighlightMatch
- components/UsersTab/: self-contained users grid with all state

Migration page no longer loads users (UsersTab not rendered there).
All 19 web tests pass. 1170 desktop tests pass.
