+++ web/src/components/common/SettingsModal.jsx
@@ -1,0 +1,22 @@
+export default function SettingsModal({ onClose }) {
+  return (
+    <div className="modal-backdrop" role="presentation" onClick={onClose}>
+      <section
+        className="modal-card utility-modal"
+        role="dialog"
+        aria-modal="true"
+        aria-labelledby="settings-modal-title"
+        onClick={(event) => event.stopPropagation()}
+      >
+        <div className="modal-header">
+          <div>
+            <h2 id="settings-modal-title">Settings</h2>
+          </div>
+          <button className="ghost-button" type="button" onClick={onClose}>Close</button>
+        </div>
+
+        <p className="status-note">No settings available.</p>
+      </section>
+    </div>
+  );
+}
